+++ crates/unftp-sbe-gcs/src/workload_identity.rs
